finished-ish encapsulating firebase functions
