+++ App.tsx
@@ -3,8 +3,7 @@ import { Provider } from 'react-redux';
 
 
 import store from './src/redux/store';
-
-import RootNavigation from './src/navigation';
+import RootNavigation from './src/screens/navigation';
 
 
 export default function App() {
