--- src/assets/constants/index.ts
@@ -1,5 +1,0 @@
-import { STATUS } from './practice_data_fields';
-import { INSTRUMENTS, LEVELS } from './profile_fields';
-
-
-export { STATUS, INSTRUMENTS, LEVELS };
