--- src/assets/constants/practice_data_fields.ts
@@ -1,4 +1,0 @@
-const STATUS = ['Not Started', 'Started But Not Completed', 'Completed'];
-
-
-export { STATUS };
