+++ src/components/headers_and_sections/profile_logo_section.tsx
@@ -6,7 +6,7 @@ import { StyleProp, ViewStyle, Alert, View, Image, Text, TouchableOpacity } from
 
 import { RootState } from '../../redux/store';
 import { setProfile } from '../../redux/actions';
-import { DataManagementAPI } from '../../services/data_management_api';
+import { DataManagementAPI } from '../../services/apis/data_management_api';
 
 import { colorPallete } from '../../assets/design_library';
