--- src/config/firebase.ts
@@ -1,25 +1,0 @@
-import Constants from 'expo-constants';
-import { getFirestore } from "firebase/firestore";
-import { initializeApp, getApp } from "firebase/app";
-import ReactNativeAsyncStorage from '@react-native-async-storage/async-storage';
-import { initializeAuth, getAuth, getReactNativePersistence } from 'firebase/auth';
-// TODO: Add other SDKs as needed
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: Constants.expoConfig?.extra?.firebaseApiKey,
-  authDomain: Constants.expoConfig?.extra?.firebaseAuthDomain,
-  projectId: Constants.expoConfig?.extra?.firebaseProjectId,
-  storageBucket: Constants.expoConfig?.extra?.firebaseStorageBucket,
-  messagingSenderId: Constants.expoConfig?.extra?.firebaseMessagingSenderId,
-  appId: Constants.expoConfig?.extra?.firebaseAppId
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const auth = initializeAuth(app, {
-  persistence: getReactNativePersistence(ReactNativeAsyncStorage)
-});
-const db = getFirestore(app);
-export { app, auth, db, getApp, getAuth };
