+++ src/services/config/firebase.ts
@@ -1,0 +1,27 @@
+import Constants from 'expo-constants';
+import ReactNativeAsyncStorage from '@react-native-async-storage/async-storage';
+
+import { initializeApp, getApp } from 'firebase/app';
+
+import { getFirestore } from 'firebase/firestore';
+import { initializeAuth, getAuth, getReactNativePersistence } from 'firebase/auth';
+
+// Add other SDKs as needed : https://firebase.google.com/docs/web/setup#available-libraries
+
+
+// Firebase configurations for 'Web Application'
+const firebase_configs = {
+    apiKey: Constants.expoConfig?.extra?.firebaseApiKey,
+    authDomain: Constants.expoConfig?.extra?.firebaseAuthDomain,
+    projectId: Constants.expoConfig?.extra?.firebaseProjectId,
+    storageBucket: Constants.expoConfig?.extra?.firebaseStorageBucket,
+    messagingSenderId: Constants.expoConfig?.extra?.firebaseMessagingSenderId,
+    appId: Constants.expoConfig?.extra?.firebaseAppId
+}
+
+// Firebase Initilization
+const app = initializeApp(firebase_configs);
+const auth = initializeAuth(app, { persistence: getReactNativePersistence(ReactNativeAsyncStorage) });
+const db = getFirestore(app);
+
+export { app, auth, db, getApp, getAuth };
